Upgrade react to 19
This included replacing enzyme with react testing library.

emoji-mart/react wasn't compatible, so we introduced a thin react wrapper (which is based on the latest implementation of emoji-mart/react).

useSelector warned about an object being returned, so we updated the equalityFn to check for diff in the object, as opposed to its reference.

React redux types for useDispatch changed, so we updated the types that we pass into it.

defaultProps was dropped by React so we replaced it with plain JS default object property values